requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
black==26.1.0
google-cloud-discoveryengine==0.17.0
python-dotenv==1.2.1